Added CircleBuilder, moved builders to different package
