+++ src/main/java/net/minestom/server/particle/shapes/builder/PolygonBuilder.java
@@ -1,5 +1,6 @@
-package net.minestom.server.particle.shapes;
+package net.minestom.server.particle.shapes.builder;
 
+import net.minestom.server.particle.shapes.ParticlePolygon;
 import net.minestom.server.utils.Position;
 import org.jetbrains.annotations.NotNull;
